[Refactor/#115] 비지니스 로직 코드 개선

+++ domain/src/main/java/com/teambrake/brake/domain/usecaseImpl/DeleteGroupUseCaseImpl.kt
@@ -1,5 +1,8 @@
 package com.teambrake.brake.domain.usecaseImpl
 
+import com.teambrake.brake.domain.model.result.BrakeResult
+import com.teambrake.brake.domain.model.result.error.DeleteGroupUseCaseError
+import com.teambrake.brake.domain.model.result.error.LocalApiCallError
 import com.teambrake.brake.domain.repository.AppGroupRepository
 import com.teambrake.brake.domain.repository.AppRepository
 import com.teambrake.brake.domain.usecase.DeleteGroupUseCase
@@ -10,15 +13,23 @@ class DeleteGroupUseCaseImpl @Inject constructor(
 	private val appGroupRepository: AppGroupRepository,
 ) : DeleteGroupUseCase {
 
-	override suspend fun invoke(
-		onError: suspend (Throwable) -> Unit,
-		groupId: Long,
-	) {
-		try {
-			appGroupRepository.deleteAppGroupByGroupId(groupId)
-			appRepository.deleteAppByParentGroupId(groupId)
-		} catch (e: Exception) {
-			onError(e)
+	override suspend fun invoke(groupId: Long): BrakeResult<Unit, DeleteGroupUseCaseError> {
+		val groupResult = appGroupRepository.deleteAppGroupByGroupId(groupId)
+
+		if (groupResult.isFailure) {
+			val e = groupResult.exceptionOrNull()
+			return BrakeResult.Error(LocalApiCallError(e ?: Exception("앱 그룹 삭제 실패")))
 		}
+
+		val appResult = appRepository.deleteAppByParentGroupId(groupId)
+
+		return appResult.fold(
+			onSuccess = {
+				BrakeResult.Success(Unit)
+			},
+			onFailure = { e ->
+				BrakeResult.Error(LocalApiCallError(e))
+			},
+		)
 	}
 }
